perbarui data

+++ kuliah/pertemuan11/detail.php
@@ -24,8 +24,8 @@ $m = query("SELECT * FROM mahasiswa WHERE id = $id");
     <li>Nama : <?= $m['nama']; ?></li>
     <li>Email : <?= $m['email']; ?></li>
     <li>Jurusan : <?= $m['jurusan']; ?></li>
-    <li><a href="ubah.php">ubah</a> | <a href="hapus.php?id=<?= $m['id']; ?>">hapus</a></li>
-    <li><a href="index.php">kembali ke daftar mahasiswa</a></li>
+    <li><a href="ubah.php">ubah</a> | <a href="hapus.php?id=<?= $m['id']; ?>" onclick="return confirm ('apakah anda yakin?');">hapus</a></li>
+    <li><a href=" index.php">kembali ke daftar mahasiswa</a></li>
   </ul>
 </body>
